Create ui component for AccountMenu
