+++ components/UI/HomePageLayout.module.css
@@ -33,7 +33,7 @@
 }
 
 .nav_message {
-    font-size: 130%;
+    font-size: 120%;
     display: flex;
     align-items: center;
     column-gap: 2vw;
