+++ components/UI/PageLayout.module.css
@@ -170,7 +170,8 @@
         font-size: 130%;
     }
 
-    .btn_user_actions {
-        margin-right: 31vw !important;
+    .mobile_left_part {
+        display: flex;
+        align-items: center;
     }
 }
