+++ components/UI/PageLayout.tsx
@@ -5,11 +5,7 @@ import Link from 'next/link';
 import MenuIcon from '@mui/icons-material/Menu';
 import CancelIcon from '@mui/icons-material/Cancel';
 import { useRouter } from 'next/router';
-import Button from '@mui/material/Button';
-import Menu from '@mui/material/Menu';
-import MenuItem from '@mui/material/MenuItem';
-import LogoutIcon from '@mui/icons-material/Logout';
-import AccountCircleIcon from '@mui/icons-material/AccountCircle';
+import { AccountSubMenu } from './AccountSubMenu';
 
 // This is Layout component for all pages in the application (exclude Home Page)
 export function PageLayout({ pageTitle, children }: { pageTitle?: string; children: ReactNode }) {
@@ -25,14 +21,6 @@ export function PageLayout({ pageTitle, children }: { pageTitle?: string; childr
     }
 
     const [openHamburgerNav, setOpenHamburgerNav] = useState(false);
-    const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-    const open = Boolean(anchorEl);
-    const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-        setAnchorEl(event.currentTarget);
-    };
-    const handleClose = () => {
-        setAnchorEl(null);
-    };
 
     const handleClickLogout = () => {
         // Logout function - Need to add here
@@ -58,57 +46,17 @@ export function PageLayout({ pageTitle, children }: { pageTitle?: string; childr
                     </div>
                 </div>
 
-                <div className={style.nav_actions}>
-                    <Button
-                        id="basic-button"
-                        aria-controls={open ? 'basic-menu' : undefined}
-                        aria-haspopup="true"
-                        aria-expanded={open ? 'true' : undefined}
-                        onClick={handleClick}
-                        sx={{
-                            minWidth: "unset",
-                        }}
-                        className={style.btn_user_actions}
-                    >
-                        <AccountCircleIcon htmlColor='#020079' fontSize='large' />
-                    </Button>
-                    <Menu
-                        id="basic-menu"
-                        anchorEl={anchorEl}
-                        open={open}
-                        onClose={handleClose}
-                        MenuListProps={{
-                            'aria-labelledby': 'basic-button',
-                        }}
-                        sx={{
-                            "ul": {
-                                padding: 0,
-                                minHeight: "unset"
-                            },
-                            "ul li": {
-                                minHeight: "unset",
-                                xs: {
-                                    padding: "3.5vw 4vw",
-                                },
-                                sm: {
-                                    padding: "0.75vw 1.5vw"
-                                }
-                            }
+                <div className={style.mobile_left_part}>
+                    <div className={style.nav_actions}>
+                        <AccountSubMenu handleClickLogout={handleClickLogout} />
+                    </div>
 
-                        }}
-                    >
-                        <MenuItem onClick={handleClickLogout} className={style.side_menu_item}>
-                            <span className={style.side_menu_label}>התנתקות</span>
-                            <span className={style.side_menu_button}><LogoutIcon sx={{ height: "fit-content" }} /></span>
-                        </MenuItem>
-                    </Menu>
-                </div>
-
-                {/** Mobile Hamburger button **/}
-                <div className={style.hamburger_btn}>
-                    <button type='button' onClick={handleClickHamburger}>
-                        <MenuIcon sx={{ fontSize: "350%" }} color='inherit' />
-                    </button>
+                    {/** Mobile Hamburger button **/}
+                    <div className={style.hamburger_btn}>
+                        <button type='button' onClick={handleClickHamburger}>
+                            <MenuIcon sx={{ fontSize: "350%" }} color='inherit' />
+                        </button>
+                    </div>
                 </div>
             </nav>
             <main className={style.main}>
